Split Draughts and Intelligence
